+++ member-api/src/main/kotlin/io/waterkite94/stalk/api/dto/response/VerifyEmailResponse.kt
@@ -1,0 +1,5 @@
+package io.waterkite94.stalk.api.dto.response
+
+data class VerifyEmailResponse(
+    val toEmail: String
+)
