--- member-api/src/main/kotlin/io/waterkite94/stalk/api/response/VerificationEmailResponse.kt
@@ -1,5 +1,0 @@
-package io.waterkite94.stalk.api.response
-
-data class VerificationEmailResponse(
-    val toEmail: String
-)
