+++ member-usecase/src/main/kotlin/io/waterkite94/stalk/usecase/usecase/CreateMemberService.kt
@@ -1,7 +1,7 @@
 package io.waterkite94.stalk.usecase.usecase
 
 import io.waterkite94.stalk.domain.model.Member
-import io.waterkite94.stalk.domain.type.RoleLevel
+import io.waterkite94.stalk.domain.vo.RoleLevel
 import io.waterkite94.stalk.exception.DuplicatedMemberException
 import io.waterkite94.stalk.security.util.SecurityUtil
 import io.waterkite94.stalk.usecase.port.AuthenticationCodePort
