--- support/security/src/main/kotlin/io/waterkite94/stalk/security/config/SecurityConfig.kt
@@ -1,12 +1,0 @@
-package io.waterkite94.stalk.security.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-import org.springframework.security.crypto.password.PasswordEncoder
-
-@Configuration
-internal class SecurityConfig {
-    @Bean
-    fun passwordEncoder(): PasswordEncoder = BCryptPasswordEncoder()
-}
